refactor(example): rewrite typescript
